moved counter to examples dir
