+++ examples/counter/cmd/countercli/commands/query.go
@@ -7,7 +7,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/commands"
 	"github.com/cosmos/cosmos-sdk/client/commands/query"
 
-	"github.com/cosmos/cosmos-sdk/docs/guide/counter/plugins/counter"
+	"github.com/cosmos/cosmos-sdk/examples/counter/plugins/counter"
 	"github.com/cosmos/cosmos-sdk/stack"
 )
