+++ examples/counter/plugins/counter/counter.go
