+++ examples/counter/plugins/counter/counter_test.go
